BPF: remove duplicate cflags and update btf flags.
With the addition of the -Wall, -Werror, -Wextra cflags to the bpf soong
module as default, they are no longer necessary in the bpf definition in
Android.bp files.

Additionally, 'btf: true'  is now set by default by the build system. To counter this and maintain the existing behavior, add 'btf: false' where necessary.

Test: treehugger
Change-Id: I82e8a316b91cd90df997f77aa7cdbf55d808d996
Signed-off-by: Neill Kapron <nkapron@google.com>