MOCK_BPF: Remove bpf mocking library
Removing the bpf mocking system because we feel that it is not useful
in properly testing bpf programs. There is currently only one real user
of it (time_in_state_test.cpp), which has significant issues in that the
tests are written in such a way that each one relies on the state of the
previous test. Also, the per CPU mocking capability is limited in its
capabilities.

Removal simplifies the transition to libbpf and reduces legacy tech
debt.

Bug: 359646531
Test: TH
Change-Id: Ic2e63ab33266e98a296ed33bddba9f3bb0029fc2
Signed-off-by: Neill Kapron <nkapron@google.com>