BpfProgs: timeInState: enable libbpf
This change ports timeInState to libbpf APIs.

This program successfully loads with BpfLoader.rs, permissions validated
by running 'ls -alh /sys/fs/bpf' on with this change set and comparing
with an unmodified cuttlefish instance.

This will open the path to deprecating the uid_sys_stats out of tree
kernel driver, as we will be able to use CO-RE to access the task
struct.

Bug: 359646531
Bug: 120442023
Bug: 329254993
Bug: 141206930
Test: Manual
Change-Id: I5229f2ac7e6663c20078f3a06f4f327f9b6b5a82
Signed-off-by: Neill Kapron <nkapron@google.com>